feat: reorganize traits to remove redundant functions from contracts (#283)
Previously, `ACLNonReentrantTrait` was a Swiss knife that combined all traits we commonly use, which had its benefits, but resulted in contracts having functions that they never use (like `setController` in `CreditFacadeV3` or `pause` in `PoolQuotaKeeperV3`), which is not really good considering that contract size limit, well, exists. This PR solves the original issue and makes separation of all functionality much cleaner.