feat: simplify `TumblerV3`'s deployment (#289)
Previously, `TumblerV3`'s constructor was accepting `pool` as argument and reading `poolQuotaKeeper` from there. This caused complications in deployments since it required a configurator-only function `setPoolQuotaKeeper` to be called between contract creations. This PR changes the constructor to accept `poolQuotaKeeper` instead of the pool, which allows to deploy contracts first and configure them later.
Note that this is the only occurrence of such issue in core contracts.